refactor(gfn): extract OAuth primitives from AuthService into auth/
Split the oversized auth.ts into focused modules (constants, oauthFlow,
deviceLogin, tokenRefresh, userInfo, helpers) while keeping AuthService
as the session orchestration owner and preserving the public export path.
