+++ opennow-stable/src/main/platforms/gfn/auth.ts
@@ -1,9 +1,6 @@
-import { createServer } from "node:http";
-import { createHash, randomBytes } from "node:crypto";
 import { access, mkdir, readFile, writeFile } from "node:fs/promises";
 import { dirname } from "node:path";
-import net from "node:net";
-import os from "node:os";
+import { randomBytes } from "node:crypto";
 
 import { shell } from "electron";
 
@@ -23,32 +20,30 @@ import type {
   StreamRegion,
   SubscriptionInfo,
 } from "@shared/gfn";
-import {
-  buildGfnLcarsHeaders,
-  buildNvidiaAuthHeaders,
-  GFN_PLAY_ORIGIN,
-  GFN_PLAY_REFERER,
-  GFN_USER_AGENT,
-} from "./clientHeaders";
+import { buildGfnLcarsHeaders, GFN_USER_AGENT } from "./clientHeaders";
 import { fetchSubscription, fetchDynamicRegions } from "./subscription";
-
-const SERVICE_URLS_ENDPOINT = "https://pcs.geforcenow.com/v1/serviceUrls";
-const TOKEN_ENDPOINT = "https://login.nvidia.com/token";
-const CLIENT_TOKEN_ENDPOINT = "https://login.nvidia.com/client_token";
-const USERINFO_ENDPOINT = "https://login.nvidia.com/userinfo";
-const AUTH_ENDPOINT = "https://login.nvidia.com/authorize";
-const DEVICE_AUTHORIZE_ENDPOINT = "https://login.nvidia.com/device/authorize";
-
-const CLIENT_ID = "ZU7sPN-miLujMD95LfOQ453IB0AtjM8sMyvgJ9wCXEQ";
-const STEAM_DECK_CLIENT_ID = "q61ddeJrVt7O90Nl-P-N7I36yctih4Ml6FyXLrb6j-U";
-const SCOPES = "openid consent email tk_client age";
-const DEFAULT_IDP_ID = "PDiAhv2kJTFeQ7WOPqiQ2tRZ7lGhR2X11dXvM4TZSxg";
-const STEAM_DECK_USER_AGENT =
-  "Mozilla/5.0 (X11; Linux x86_64; Steam Deck) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36";
-
-const REDIRECT_PORTS = [2259, 6460, 7119, 8870, 9096];
-const TOKEN_REFRESH_WINDOW_MS = 10 * 60 * 1000;
-const CLIENT_TOKEN_REFRESH_WINDOW_MS = 5 * 60 * 1000;
+import {
+  CLIENT_TOKEN_REFRESH_WINDOW_MS,
+  DEFAULT_IDP_ID,
+  SERVICE_URLS_ENDPOINT,
+  TOKEN_REFRESH_WINDOW_MS,
+} from "./auth/constants";
+import { isExpired, isNearExpiry } from "./auth/helpers";
+import {
+  buildAuthUrl,
+  exchangeAuthorizationCode,
+  findAvailablePort,
+  generatePkce,
+  waitForAuthorizationCode,
+} from "./auth/oauthFlow";
+import { exchangeDeviceCode, requestDeviceAuthorization } from "./auth/deviceLogin";
+import {
+  mergeTokenSnapshot,
+  refreshAuthTokens,
+  refreshWithClientToken,
+  requestClientToken,
+} from "./auth/tokenRefresh";
+import { fetchUserInfo } from "./auth/userInfo";
 
 interface PersistedAuthState {
   sessions: AuthSession[];
@@ -69,33 +64,6 @@ interface ServiceUrlsResponse {
       loginProviderPriority?: number;
     }>;
   };
-}
-
-interface TokenResponse {
-  access_token: string;
-  refresh_token?: string;
-  id_token?: string;
-  client_token?: string;
-  expires_in?: number;
-}
-
-interface ClientTokenResponse {
-  client_token: string;
-  expires_in?: number;
-}
-
-interface DeviceAuthorizationResponse {
-  device_code?: string;
-  user_code?: string;
-  verification_uri?: string;
-  verification_uri_complete?: string;
-  expires_in?: number;
-  interval?: number;
-}
-
-interface DeviceTokenErrorResponse {
-  error?: string;
-  error_description?: string;
 }
 
 interface ServerInfoResponse {
@@ -130,451 +98,6 @@ function normalizeProvider(provider: LoginProvider): LoginProvider {
     streamingServiceUrl: provider.streamingServiceUrl.endsWith("/")
       ? provider.streamingServiceUrl
       : `${provider.streamingServiceUrl}/`,
-  };
-}
-
-function decodeBase64Url(value: string): string {
-  const normalized = value.replace(/-/g, "+").replace(/_/g, "/");
-  const padding = normalized.length % 4;
-  const padded = padding === 0 ? normalized : `${normalized}${"=".repeat(4 - padding)}`;
-  return Buffer.from(padded, "base64").toString("utf8");
-}
-
-function parseJwtPayload<T>(token: string): T | null {
-  const parts = token.split(".");
-  if (parts.length !== 3) {
-    return null;
-  }
-  try {
-    const payload = decodeBase64Url(parts[1]);
-    return JSON.parse(payload) as T;
-  } catch {
-    return null;
-  }
-}
-
-function toExpiresAt(expiresInSeconds: number | undefined, defaultSeconds = 86400): number {
-  return Date.now() + (expiresInSeconds ?? defaultSeconds) * 1000;
-}
-
-function isExpired(expiresAt: number | undefined): boolean {
-  if (!expiresAt) {
-    return true;
-  }
-  return expiresAt <= Date.now();
-}
-
-function isNearExpiry(expiresAt: number | undefined, windowMs: number): boolean {
-  if (!expiresAt) {
-    return true;
-  }
-  return expiresAt - Date.now() < windowMs;
-}
-
-function generateDeviceId(): string {
-  const host = os.hostname();
-  const username = os.userInfo().username;
-  return createHash("sha256").update(`${host}:${username}:opennow-stable`).digest("hex");
-}
-
-function generatePkce(): { verifier: string; challenge: string } {
-  const verifier = randomBytes(64)
-    .toString("base64")
-    .replace(/\+/g, "-")
-    .replace(/\//g, "_")
-    .replace(/=+$/g, "")
-    .slice(0, 86);
-
-  const challenge = createHash("sha256")
-    .update(verifier)
-    .digest("base64")
-    .replace(/\+/g, "-")
-    .replace(/\//g, "_")
-    .replace(/=+$/g, "");
-
-  return { verifier, challenge };
-}
-
-function buildAuthHeadersForClient(
-  authClientId = CLIENT_ID,
-  options: {
-    bearerToken?: string;
-    accept?: string;
-    contentType?: string;
-    includeReferer?: boolean;
-  } = {},
-): Record<string, string> {
-  if (authClientId !== STEAM_DECK_CLIENT_ID) {
-    return buildNvidiaAuthHeaders(options);
-  }
-
-  const headers: Record<string, string> = {
-    Accept: options.accept ?? "application/json, text/plain, */*",
-    Origin: GFN_PLAY_ORIGIN,
-    Referer: GFN_PLAY_REFERER,
-    "User-Agent": STEAM_DECK_USER_AGENT,
-  };
-
-  if (options.bearerToken !== undefined) {
-    headers.Authorization = `Bearer ${options.bearerToken}`;
-  }
-  if (options.contentType) {
-    headers["Content-Type"] = options.contentType;
-  }
-
-  return headers;
-}
-
-function buildAuthUrl(provider: LoginProvider, challenge: string, port: number): string {
-  const redirectUri = `http://localhost:${port}`;
-  const nonce = randomBytes(16).toString("hex");
-  const params = new URLSearchParams({
-    response_type: "code",
-    device_id: generateDeviceId(),
-    scope: SCOPES,
-    client_id: CLIENT_ID,
-    redirect_uri: redirectUri,
-    ui_locales: "en_US",
-    nonce,
-    prompt: "select_account",
-    code_challenge: challenge,
-    code_challenge_method: "S256",
-    idp_id: provider.idpId,
-  });
-  return `${AUTH_ENDPOINT}?${params.toString()}`;
-}
-
-async function isPortAvailable(port: number): Promise<boolean> {
-  return new Promise((resolve) => {
-    const server = net.createServer();
-    server.once("error", () => resolve(false));
-    server.once("listening", () => {
-      server.close(() => resolve(true));
-    });
-    server.listen(port, "127.0.0.1");
-  });
-}
-
-async function findAvailablePort(): Promise<number> {
-  for (const port of REDIRECT_PORTS) {
-    if (await isPortAvailable(port)) {
-      return port;
-    }
-  }
-
-  throw new Error("No available OAuth callback ports");
-}
-
-async function waitForAuthorizationCode(port: number, timeoutMs: number): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const server = createServer((request, response) => {
-      const url = new URL(request.url ?? "/", `http://localhost:${port}`);
-      const code = url.searchParams.get("code");
-      const error = url.searchParams.get("error");
-
-      const html = `<!doctype html><html><head><meta charset="utf-8"><title>OpenNOW Login</title></head><body style="font-family:Segoe UI,Arial,sans-serif;background:#0b1220;color:#dbe7ff;display:flex;justify-content:center;align-items:center;height:100vh"><div style="background:#111a2c;padding:24px 28px;border:1px solid #30425f;border-radius:12px;max-width:460px"><h2 style="margin-top:0">OpenNOW Login</h2><p>${
-        code
-          ? "Login complete. You can close this window and return to OpenNOW Stable."
-          : "Login failed or was cancelled. You can close this window and return to OpenNOW Stable."
-      }</p></div></body></html>`;
-
-      response.statusCode = 200;
-      response.setHeader("Content-Type", "text/html; charset=utf-8");
-      response.end(html);
-
-      server.close(() => {
-        if (code) {
-          resolve(code);
-          return;
-        }
-        reject(new Error(error ?? "Authorization failed"));
-      });
-    });
-
-    server.listen(port, "127.0.0.1", () => {
-      const timer = setTimeout(() => {
-        server.close(() => reject(new Error("Timed out waiting for OAuth callback")));
-      }, timeoutMs);
-
-      server.once("close", () => clearTimeout(timer));
-    });
-  });
-}
-
-async function exchangeAuthorizationCode(code: string, verifier: string, port: number): Promise<AuthTokens> {
-  const body = new URLSearchParams({
-    grant_type: "authorization_code",
-    code,
-    redirect_uri: `http://localhost:${port}`,
-    code_verifier: verifier,
-  });
-
-  const response = await fetch(TOKEN_ENDPOINT, {
-    method: "POST",
-    headers: buildAuthHeadersForClient(CLIENT_ID, {
-      contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-      includeReferer: true,
-    }),
-    body,
-  });
-
-  if (!response.ok) {
-    const text = await response.text();
-    throw new Error(`Token exchange failed (${response.status}): ${text.slice(0, 400)}`);
-  }
-
-  const payload = (await response.json()) as TokenResponse;
-  return {
-    accessToken: payload.access_token,
-    refreshToken: payload.refresh_token,
-    idToken: payload.id_token,
-    expiresAt: toExpiresAt(payload.expires_in),
-    authClientId: CLIENT_ID,
-  };
-}
-
-async function requestDeviceAuthorization(
-  provider: LoginProvider,
-): Promise<Omit<AuthDeviceLoginChallenge, "attemptId">> {
-  const deviceId = generateDeviceId();
-  const body = new URLSearchParams({
-    client_id: STEAM_DECK_CLIENT_ID,
-    scope: SCOPES,
-    device_id: deviceId,
-    display_name: "OpenNOW",
-    idp_id: provider.idpId,
-  });
-
-  const response = await fetch(DEVICE_AUTHORIZE_ENDPOINT, {
-    method: "POST",
-    headers: {
-      ...buildAuthHeadersForClient(STEAM_DECK_CLIENT_ID, {
-        contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-      }),
-      "x-device-id": deviceId,
-      "nv-client-id": STEAM_DECK_CLIENT_ID,
-      "nv-client-streamer": "WEBRTC",
-      "nv-client-type": "BROWSER",
-      "nv-client-platform-name": "browser",
-      "nv-browser-type": "CHROME",
-      "nv-device-os": "STEAMOS",
-      "nv-device-type": "CONSOLE",
-      "nv-device-model": "STEAMDECK",
-      "nv-device-make": "VALVE",
-    },
-    body,
-  });
-
-  if (!response.ok) {
-    const text = await response.text();
-    throw new Error(`Device authorization failed (${response.status}): ${text.slice(0, 400)}`);
-  }
-
-  const payload = (await response.json()) as DeviceAuthorizationResponse;
-  if (
-    !payload.device_code ||
-    !payload.user_code ||
-    !payload.verification_uri ||
-    !payload.verification_uri_complete
-  ) {
-    throw new Error("Device authorization response did not include QR login data");
-  }
-
-  return {
-    deviceCode: payload.device_code,
-    userCode: payload.user_code,
-    verificationUri: payload.verification_uri,
-    verificationUriComplete: payload.verification_uri_complete,
-    expiresAt: toExpiresAt(payload.expires_in, 600),
-    intervalSeconds: Math.max(1, payload.interval ?? 5),
-  };
-}
-
-async function exchangeDeviceCode(deviceCode: string): Promise<AuthTokens | DeviceTokenErrorResponse> {
-  const body = new URLSearchParams({
-    grant_type: "urn:ietf:params:oauth:grant-type:device_code",
-    device_code: deviceCode,
-    client_id: STEAM_DECK_CLIENT_ID,
-  });
-
-  const response = await fetch(TOKEN_ENDPOINT, {
-    method: "POST",
-    headers: buildAuthHeadersForClient(STEAM_DECK_CLIENT_ID, {
-      contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-    }),
-    body,
-  });
-
-  const payload = (await response.json().catch(() => null)) as TokenResponse | DeviceTokenErrorResponse | null;
-  if (!response.ok) {
-    return payload && typeof payload === "object"
-      ? payload as DeviceTokenErrorResponse
-      : { error: "device_token_exchange_failed", error_description: `Device token exchange failed (${response.status})` };
-  }
-
-  const tokenPayload = payload as TokenResponse | null;
-  if (!tokenPayload?.access_token) {
-    return { error: "invalid_token_response", error_description: "Device token response did not include access_token" };
-  }
-
-  return {
-    accessToken: tokenPayload.access_token,
-    refreshToken: tokenPayload.refresh_token,
-    idToken: tokenPayload.id_token,
-    expiresAt: toExpiresAt(tokenPayload.expires_in),
-    authClientId: STEAM_DECK_CLIENT_ID,
-    clientToken: tokenPayload.client_token,
-  };
-}
-
-async function refreshAuthTokens(refreshToken: string, authClientId = CLIENT_ID): Promise<AuthTokens> {
-  const body = new URLSearchParams({
-    grant_type: "refresh_token",
-    refresh_token: refreshToken,
-    client_id: authClientId,
-  });
-
-  const response = await fetch(TOKEN_ENDPOINT, {
-    method: "POST",
-    headers: buildAuthHeadersForClient(authClientId, {
-      contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-    }),
-    body,
-  });
-
-  if (!response.ok) {
-    const text = await response.text();
-    throw new Error(`Token refresh failed (${response.status}): ${text.slice(0, 400)}`);
-  }
-
-  const payload = (await response.json()) as TokenResponse;
-  return {
-    accessToken: payload.access_token,
-    refreshToken: payload.refresh_token ?? refreshToken,
-    idToken: payload.id_token,
-    expiresAt: toExpiresAt(payload.expires_in),
-    authClientId,
-  };
-}
-
-async function requestClientToken(accessToken: string, authClientId = CLIENT_ID): Promise<{
-  token: string;
-  expiresAt: number;
-  lifetimeMs: number;
-}> {
-  const response = await fetch(CLIENT_TOKEN_ENDPOINT, {
-    headers: buildAuthHeadersForClient(authClientId, { bearerToken: accessToken }),
-  });
-
-  if (!response.ok) {
-    const text = await response.text();
-    throw new Error(`Client token request failed (${response.status}): ${text.slice(0, 400)}`);
-  }
-
-  const payload = (await response.json()) as ClientTokenResponse;
-  const expiresAt = toExpiresAt(payload.expires_in);
-  return {
-    token: payload.client_token,
-    expiresAt,
-    lifetimeMs: Math.max(0, expiresAt - Date.now()),
-  };
-}
-
-async function refreshWithClientToken(clientToken: string, userId: string, authClientId = CLIENT_ID): Promise<TokenResponse> {
-  const body = new URLSearchParams({
-    grant_type: "urn:ietf:params:oauth:grant-type:client_token",
-    client_token: clientToken,
-    client_id: authClientId,
-    sub: userId,
-  });
-
-  const response = await fetch(TOKEN_ENDPOINT, {
-    method: "POST",
-    headers: buildAuthHeadersForClient(authClientId, {
-      contentType: "application/x-www-form-urlencoded; charset=UTF-8",
-    }),
-    body,
-  });
-
-  if (!response.ok) {
-    const text = await response.text();
-    throw new Error(`Client-token refresh failed (${response.status}): ${text.slice(0, 400)}`);
-  }
-
-  return (await response.json()) as TokenResponse;
-}
-
-function mergeTokenSnapshot(base: AuthTokens, refreshed: TokenResponse): AuthTokens {
-  return {
-    accessToken: refreshed.access_token,
-    refreshToken: refreshed.refresh_token ?? base.refreshToken,
-    idToken: refreshed.id_token,
-    expiresAt: toExpiresAt(refreshed.expires_in),
-    authClientId: base.authClientId ?? CLIENT_ID,
-    clientToken: refreshed.client_token ?? base.clientToken,
-    clientTokenExpiresAt: base.clientTokenExpiresAt,
-    clientTokenLifetimeMs: base.clientTokenLifetimeMs,
-  };
-}
-
-function gravatarUrl(email: string, size = 80): string {
-  const normalized = email.trim().toLowerCase();
-  const hash = createHash("md5").update(normalized).digest("hex");
-  return `https://www.gravatar.com/avatar/${hash}?s=${size}&d=identicon`;
-}
-
-async function fetchUserInfo(tokens: AuthTokens): Promise<AuthUser> {
-  const jwtToken = tokens.idToken ?? tokens.accessToken;
-  const parsed = parseJwtPayload<{
-    sub?: string;
-    email?: string;
-    preferred_username?: string;
-    gfn_tier?: string;
-    picture?: string;
-  }>(jwtToken);
-
-  if (parsed?.sub) {
-    const emailFromToken = parsed.email;
-    const pictureFromToken = parsed.picture;
-    if (emailFromToken || pictureFromToken) {
-      const avatar = pictureFromToken ?? (emailFromToken ? gravatarUrl(emailFromToken) : undefined);
-      return {
-        userId: parsed.sub,
-        displayName: parsed.preferred_username ?? emailFromToken?.split("@")[0] ?? "User",
-        email: emailFromToken,
-        avatarUrl: avatar,
-        membershipTier: parsed.gfn_tier ?? "FREE",
-      };
-    }
-  }
-
-  const response = await fetch(USERINFO_ENDPOINT, {
-    headers: buildAuthHeadersForClient(tokens.authClientId, {
-      bearerToken: tokens.accessToken,
-      accept: "application/json",
-    }),
-  });
-
-  if (!response.ok) {
-    throw new Error(`User info failed (${response.status})`);
-  }
-
-  const payload = (await response.json()) as {
-    sub: string;
-    preferred_username?: string;
-    email?: string;
-    picture?: string;
-  };
-
-  const email = payload.email;
-  const avatar = payload.picture ?? (email ? gravatarUrl(email) : undefined);
-
-  return {
-    userId: payload.sub,
-    displayName: payload.preferred_username ?? email?.split("@")[0] ?? "User",
-    email,
-    avatarUrl: avatar,
-    membershipTier: "FREE",
   };
 }
