+++ opennow-stable/src/main/platforms/gfn/auth/constants.ts
@@ -1,0 +1,17 @@
+export const SERVICE_URLS_ENDPOINT = "https://pcs.geforcenow.com/v1/serviceUrls";
+export const TOKEN_ENDPOINT = "https://login.nvidia.com/token";
+export const CLIENT_TOKEN_ENDPOINT = "https://login.nvidia.com/client_token";
+export const USERINFO_ENDPOINT = "https://login.nvidia.com/userinfo";
+export const AUTH_ENDPOINT = "https://login.nvidia.com/authorize";
+export const DEVICE_AUTHORIZE_ENDPOINT = "https://login.nvidia.com/device/authorize";
+
+export const CLIENT_ID = "ZU7sPN-miLujMD95LfOQ453IB0AtjM8sMyvgJ9wCXEQ";
+export const STEAM_DECK_CLIENT_ID = "q61ddeJrVt7O90Nl-P-N7I36yctih4Ml6FyXLrb6j-U";
+export const SCOPES = "openid consent email tk_client age";
+export const DEFAULT_IDP_ID = "PDiAhv2kJTFeQ7WOPqiQ2tRZ7lGhR2X11dXvM4TZSxg";
+export const STEAM_DECK_USER_AGENT =
+  "Mozilla/5.0 (X11; Linux x86_64; Steam Deck) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/128.0.0.0 Safari/537.36";
+
+export const REDIRECT_PORTS = [2259, 6460, 7119, 8870, 9096];
+export const TOKEN_REFRESH_WINDOW_MS = 10 * 60 * 1000;
+export const CLIENT_TOKEN_REFRESH_WINDOW_MS = 5 * 60 * 1000;
